pyproject.toml
[project]
name = "nonebot-plugin-arkgacha"
version = "0.8.0"
description = "Plugin for Arknights gacha, support multi platform"
authors = [
    {name = "RF-Tar-Railt", email = "rf_tar_railt@qq.com"},
]
dependencies = [
    "nonebot2>=2.3.0",
    "nonebot-plugin-alconna>=0.50.2",
    "arknights-toolkit>=0.7.5",
    "nonebot-plugin-localstore>=0.6.0",
    "nonebot-plugin-apscheduler>=0.4.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[project.entry-points.nb_scripts]
arkgacha = "nonebot_plugin_arkgacha.__main__:main"

[tool.pdm]

[tool.pdm.build]
includes = ["nonebot_plugin_arkgacha"]
[tool.pdm.dev-dependencies]
dev = [
    "nonebot-adapter-qq>=1.4.4",
    "nonebot2[httpx,websockets]>=2.3.0",
]
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
